requirements.txt
pandas
matplotlib
findpeaks
plotly
tabulate
rich
numpy
scipy
seaborn
colorama
pyyaml
pyvista
holoviews
datashader
bokeh
vtk
PyQt5
kaleido
blessed
rich